+++ apps/web/.eslintrc.cjs
@@ -1,4 +1,4 @@
 module.exports = {
   root: true,
-  extends: [require.resolve('@world-quest/config/eslint/base.cjs'), 'next/core-web-vitals'],
+  extends: ['next/core-web-vitals'],
 };